fixed ui error bug
The error is from having StandaloneInputModule in the scene, which is the UI input module for the old input system. If you select the EventSystem GameObject, you should see a "Replace with InputSystemUIInputModule" button in the inspector for the StandaloneInputModule component. Clicking that button and saving the scene should fix the error.